+++ packages/widgetbook_test/lib/widgetbook_test.dart
@@ -1,1 +1,1 @@
-
+export 'src/test.dart' hide outputDir;