+++ sandbox/test/widgetbook_test.dart
@@ -1,0 +1,6 @@
+import 'package:sandbox/widgetbook.config.dart';
+import 'package:widgetbook_test/widgetbook_test.dart';
+
+Future<void> main() async {
+  await testWidgetbook(config);
+}