+++ packages/widgetbook_test/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- **FEAT**: Add `testWidgetbook` function. ([#1610](https://github.com/widgetbook/widgetbook/pull/1610))
+
 ## 0.0.0
 
 - Empty release.